+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 
 - **plugin-text**. Fix positioning of text controls on touch devices
+- **plugin-text**. Split plugin when pasting multiple blocks
 
 ## [0.6.2](https://github.com/edtr-io/edtr-io/compare/v0.6.1..v0.6.2) - June 19, 2019